pyproject.toml
[project]
name = "work-clothes-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "django>=6.0",
    "django-allauth>=65.13.1",
    "django-htmx>=1.27.0",
    "hypothesis>=6.148.8",
    "pillow>=12.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
]
